+++ lib/dashboard/home/widgets/transactions_slim_view.dart
@@ -2,9 +2,9 @@ import 'package:auto_size_text/auto_size_text.dart';
 import 'package:flutter/material.dart';
 import 'package:genius_api/models/transaction.dart';
 import 'package:genius_wallet/dashboard/home/widgets/transaction_filters.dart';
-import 'package:genius_wallet/dashboard/home/widgets/transaction_item.dart';
-import 'package:genius_wallet/dashboard/home/widgets/transaction_purchased_item.dart';
-import 'package:genius_wallet/dashboard/home/widgets/transaction_swapped_item.dart';
+import 'package:genius_wallet/dashboard/transactions/transaction_item.dart';
+import 'package:genius_wallet/dashboard/transactions/transaction_purchased_item.dart';
+import 'package:genius_wallet/dashboard/transactions/transaction_swapped_item.dart';
 import 'package:genius_wallet/theme/genius_wallet_colors.dart';
 import 'package:intl/intl.dart';
