+++ lib/dashboard/transactions/transaction_item.dart
@@ -19,8 +19,15 @@ class TransactionItem extends StatelessWidget {
   Widget build(BuildContext context) {
     final isSent = tx.transactionDirection == TransactionDirection.sent;
     final label = isSent ? "Sent" : "Received";
-    final amount =
-        "${isSent ? '-' : '+'} ${formatAmount(tx.recipients.first.amount)} ${tx.coinSymbol}";
+    String amount;
+
+    try {
+      amount =
+          "${isSent ? '-' : '+'} ${formatAmount(tx.recipients.first.amount)} ${tx.coinSymbol}";
+    } on StateError {
+      amount = "";
+    }
+
     final arrowIcon = isSent ? Icons.arrow_forward : Icons.arrow_downward;
     final arrowBgColor = isSent ? Colors.lightBlueAccent : Colors.greenAccent;
