+++ lib/dashboard/transactions/transactions_screen.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:genius_api/types/wallet_type.dart';
-import 'package:genius_wallet/dashboard/home/widgets/sgnus_transactions_screen.dart';
+import 'package:genius_wallet/dashboard/transactions/transactions_scren.dart';
 
 import 'package:genius_wallet/dashboard/transactions/view/transactions_stream.dart';
 import 'package:genius_wallet/wallets/cubit/wallet_details_cubit.dart';
