+++ packages/genius_api/lib/src/genius_api.dart
@@ -136,7 +136,7 @@ class GeniusApi {
         .join();
 
     final privateKeyAsPtr = privateKeyAsStr.toNativeUtf8();
-    debugPrint('Json File Path: ${jsonFilePath}');
+    debugPrint('Json File Path: $jsonFilePath');
     final retVal = ffiBridgePrebuilt.gns_lib.GeniusSDKInit(
         basePathPtr.cast(), privateKeyAsPtr.cast(), true, true, 41001);
